stripe first iteration done (i hope)

+++ Database/python_data/schema/gym_history.py
@@ -11,4 +11,4 @@ class GymHistoryCreate(SeedModel):
     members_churned: int
     members_gained: int
     members_retained: int
-    revenue: float
+    revenue: int
